requirements.txt
anthropic>=0.40.0
apscheduler>=3.10.4
pillow>=10.4.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
requests>=2.32.3
